+++ src/main/java/com/company/services/JabatanService.java
@@ -4,6 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.company.models.entities.Jabatan;
 import com.company.models.repository.JabatanRepo;
+
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 
 @Service
@@ -26,7 +28,9 @@ public class JabatanService {
     }
 
     public Jabatan findOne(Integer id) {
-        return jabatanRepo.findById(id).get();
+        return jabatanRepo.findById(id).orElseThrow(() ->
+            new EntityNotFoundException("Absensi with ID " + id + " not found")
+        );
     }
 
     public Iterable<Jabatan> findAll() {
@@ -34,6 +38,10 @@ public class JabatanService {
     }
 
     public void removeOne(Integer id) {
+        if(jabatanRepo.existsById(id)) {
+            throw new EntityNotFoundException("Absensi with ID " + id + " not found");
+        }
+
         jabatanRepo.deleteById(id);
     }
 }
